fix: gendummydata: optdepends description must be stored in own db field
Previously we generated an optdepends description but were applying it
to the depends name. This created broken optdepends references by
accident, fix this by using the appropriate db field in most of the
cases, and include some probability for not using any optdepends
description which allows to test for all possible code paths.