feat: gendummydata: provide and depend on official packages
This greatly helps to evaluate all code path depending on official
packages. To do so, we generate AUR packages that provide official
packages as well as generate AUR packages that depend on official
packages. Official packages are chosen with regular pkgname as well as
meta provides that are fulfilled by multiple official packages.