feat: gendummydata: generate split-packages sharing the same pkgbase
This allows to test missing code paths that depend on split packages, or
resolve pkgbase from split packages. To achieve this, we have a small
likehood that a package is a split package. We generate a set of
pkgnames like before, but now we either use an already generated package
as base which makes up a split-package, or just use ourselves as the
base.