fix: cron: force remove leftover alpm sync lock files during pod startup
This may happen when the pod gets killed during aurblup alpm repo sync
where the lock file is left over. Once the pod is booted again, the
leftover lock file will lead to the cron process not being able to sync
the repo and repeatedly logging the failure in the pod logs.
It's fine to remove the lock forcefully when we startup the pod, as
before exec there is no process currently syncing.