perf: packages: reduce queries for package relations
We already query all package relations and put them in a dict. Lets
simply use this dict and assign the values to the context instead of
doing another round of queries hitting the database. On top, also avoid
querying these individual relations multiplied by two, by having the
templates do a count() call, which is equally expensive as the
subsequent query for the actual data.

This alone removes 8 queries that are hitting the database for no
reason.