perf: packages: avoid count query on comments template
We don't really need to query the count here, we already queries the
total count when making the context for packages and pkgbase views.
Using this total count is sufficient, as we only guard the html elements
for the case where there are no comments at all. If there are any
comments, we can as well just use this as boolean expression to show the
elements, independent from the current pagination.

This reduces one paged query on a semi expensive table.