perf: packages: eager join the type for package dependencies
Using lazy join when accessing the dependency-type here is diminishing
to the performance for multiple reasons. First, we basically always want
to access the actual dependency type information from the mapping table.
If we do not do this, the lazy join will lead to 4 individual queries
going back and forth between the application and the database for no
reason, which takes a noticeable delay. And secondly even if we'd not
need this, eagerly joining on the table with 4 rows is quite cheap
_during_ the query for the package_dependency without an extra round
between app and database.

This removes 4 individual queries hitting the database backend when
selecting multiple package dependencies, which is always the case.