perf: packages: lookup aur dependencies from pre fetched list
Move the aur dependencies lookup code to the utility module for cleaner
and reusable code. Also pass the dependency names to lookup from the
already fetched list of dependencies in the packages endpoint instead of
calling the relations API which effectively issues a fresh query.

This reduces one additional package_dependency query which we already
obtained, plus is a preparation for successive optimizations.