chore(pre-commit): switch to ruff for linting
Ruff supports most of the flake8 rules aurweb uses and can replace
isort. The only two rules not supported are W503/W504 which are excluded
anyway so not an issue.

Ruff formats imports slightly different where `foo as bar` are on a
separate line from other imports from the same file.